requirements.txt
python-documentcloud==3.6.0
requests
azure-ai-formrecognizer==3.2.0
